test: fix test error
